+++ src/hooks/form/useModalForm.ts
@@ -30,8 +30,6 @@ const useModalForm = () => {
   });
 
   function onSubmit(values: z.infer<typeof formSchema>) {
-    // Do something with the form values.
-    // ✅ This will be type-safe and validated.
     console.log(values);
   }
